Jenkinsfile̕ҏW
rhAÓIð́A[MA[NXy[X̃N[

+++ calc/inc/polishnotation.h
@@ -17,6 +17,6 @@ double calcFormula(char,
                    double,
                    double);
 void toPolishNotation(char*);
-double calcPolishNotation(char*);
+int calcPolishNotation(char*);
 
 #endif
